feat: Complete SHI v2 transformation (Sprints 2-5)
Sprint 2 - Graph Intelligence:
- Node2Vec embeddings for wallet similarity detection
- Dynamic network metrics (modularity, density, centralization)
- Isolation Forest anomaly detection
- Sybil cluster identification

Sprint 3 - Real-time Monitoring:
- WalletWatcher service for large movements
- Alert engine (whale_movement, regime_change, anomaly_spike)
- Profile evolution tracker
- Telegram commands: /watch, /unwatch, /alerts, /profile
- Push notification system

Sprint 4 - User Intelligence & Explainability:
- SHAP-based risk score explanations
- Natural language narrative generation
- Capital flow forecasting with confidence intervals
- Dashboard-ready data structures (Pydantic v2)
- Telegram commands: /explain, /forecast

Sprint 5 - Behavioral Intelligence & API:
- Wallet action sequence modeling and encoding
- Pre-dump signature detection
- Bayesian risk estimation with Beta posteriors
- REST API (8 endpoints with FastAPI)
- Production hardening: 267 tests passing, 0 warnings
- Telegram commands: /sequence, /belief

Test results: 267 passed, 26 skipped, 0 warnings
Mypy: 51 errors (down from 93)

+++ src/models/correlation.py
@@ -19,7 +19,6 @@ from typing import Sequence
 import numpy as np
 import structlog
 
-from ..clustering.archetypes import Archetype
 
 logger = structlog.get_logger()
 
@@ -252,8 +251,8 @@ def detect_coordinated_behavior(
         return 0.0
 
     # Compute coefficient of variation
-    mean_interval = np.mean(intervals)
-    std_interval = np.std(intervals)
+    mean_interval = float(np.mean(intervals))
+    std_interval = float(np.std(intervals))
 
     if mean_interval == 0:
         return 1.0  # All entries at same time = perfect sync
@@ -262,7 +261,7 @@ def detect_coordinated_behavior(
 
     # Lower CV = more synchronized
     # Convert to [0, 1] score where 1 = highly synchronized
-    window_seconds = temporal_window_hours * 3600
+    window_seconds = float(temporal_window_hours * 3600)
 
     if mean_interval < window_seconds:
         # Entries are within the window
